all bugs fixed
